Renamed ui's Editable to EditableUI.
