--- tests/_utils/ui/editable/inline/inlineeditable.js
@@ -1,11 +1,0 @@
-/**
- * @license Copyright (c) 2003-2016, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md.
- */
-
-'use strict';
-
-import Editable from '/ckeditor5/ui/editable/editable.js';
-
-export default class InlineEditable extends Editable {
-}
